llm_service/requirements.txt
grpcio
grpcio-tools
protobuf
openai
python-dotenv==1.1.0
